Adapted to work with Arduino UNO
Made modifications to use less RAM. Once running on the UNO,
the temperature measurements were inverted. Reworked the To algorithm
to calculate the actual observed temperature.

This project also includes the addition of an LCD screen.
I will eventually move that into a separate sketch and leave the MLX90621
library as a stand alone